% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cubist-package.R
\docType{package}
\name{Cubist-package}
\alias{Cubist}
\alias{Cubist-package}
\title{Rule-based regression modeling}
\description{
Cubist implements Quinlan's rule-based regression model. It builds
predictive models by growing a tree structure and collapsing each path into
a rule, then fitting a linear regression model for each rule based on the
data subset it covers. Predictions can optionally be improved using boosting
via committees (ensembles of rule sets) and nearest-neighbor adjustments
from training instances.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://topepo.github.io/Cubist/}
  \item \url{https://github.com/topepo/Cubist}
  \item Report bugs at \url{https://github.com/topepo/Cubist/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{mxkuhn@gmail.com}

Authors:
\itemize{
  \item Ross Quinlan (Author of imported C code)
}

Other contributors:
\itemize{
  \item Steve Weston [contributor]
  \item Chris Keefer [contributor]
  \item Nathan Coulter [contributor]
  \item Rulequest Research Pty Ltd. (Copyright holder of imported C code) [copyright holder]
}

}
\keyword{internal}
