% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuinlanAttributes.R
\name{QuinlanAttributes}
\alias{QuinlanAttributes}
\title{Convert Data to Text Format}
\usage{
QuinlanAttributes(object, ...)
}
\arguments{
\item{object}{An object to translate to the format expected by
the Quinlan C code.}
}
\value{
A text representation of the data.
}
\description{
This class converts different types of data to a format
required for \href{http://www.rulequest.com/cubist-unix.html}{Cubist}.
}
\keyword{internal}
