% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rules.R
\name{tidy_rules.cubist}
\alias{tidy_rules.cubist}
\title{Obtain rules as a tidy tibble from a cubist model}
\usage{
\method{tidy_rules}{cubist}(object, ...)
}
\arguments{
\item{object}{Fitted model object with rules}

\item{...}{Other arguments (currently unused)}
}
\value{
A tibble where each row corresponds to a rule. The columns are:
support, mean, min, max, error, lhs, rhs and committees
}
\description{
Each row corresponds to a rule. A rule can be copied into
\code{dplyr::filter} to filter the observations corresponding to a rule
}
\examples{
data("attrition", package = "rsample")
attrition <- tibble::as_tibble(attrition)
# lets predict monthly income
cubist_model <- 
  Cubist::cubist(
    x   = attrition \%>\% dplyr::select(-MonthlyIncome, -Attrition)
    , y = attrition \%>\% dplyr::select(MonthlyIncome) \%>\% unlist()
    )
summary(cubist_model)
tidy_rules(cubist_model)

cubist_model_commitees <- 
  Cubist::cubist(x   = attrition \%>\% dplyr::select(-MonthlyIncome, -Attrition)
                 , y = attrition \%>\% dplyr::select(MonthlyIncome) \%>\% unlist()
                 , committees = 7
                 )
summary(cubist_model_commitees)
tidy_rules(cubist_model_commitees)
}
