% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rules.R
\name{tidy_rules}
\alias{tidy_rules}
\title{Obtain rules as a tidy tibble}
\usage{
tidy_rules(object, ...)
}
\arguments{
\item{object}{Fitted model object with rules}

\item{...}{Other arguments (currently unused)}
}
\value{
A tibble where each row corresponds to a rule
}
\description{
Each row corresponds to a rule. A rule can be copied into
\code{dplyr::filter} to filter the observations corresponding to a rule
}
\author{
Srikanth KS, \email{sri.teach@gmail.com}
}
