\name{Cubist-package}
\alias{Cubist-package}
\alias{Cubist}
\docType{package}
\title{
Rule based models
}
\description{
More about what it does (maybe more than one line)

 A concise (1-5 lines) description of the package 

}
\details{
\tabular{ll}{
Package: \tab Cubist\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-25\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
 An overview of how to use the package, including the most important functions 
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>

 The author and/or maintainer of the package 

}
\references{

 Literature or other references for background information 

}
\keyword{ package }
\seealso{
\code{\link{cubist}}
}
\examples{
# simple examples of the most important functions
}
