% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rules.R
\name{tidy_rules.cubist}
\alias{tidy_rules.cubist}
\title{Obtain rules as a tidy tibble from a cubist model}
\usage{
\method{tidy_rules}{cubist}(object, ...)
}
\arguments{
\item{object}{Fitted model object with rules}

\item{...}{Other arguments (currently unused)}
}
\value{
A tibble where each row corresponds to a rule. The columns are:
support, mean, min, max, error, lhs, rhs and committee
}
\description{
Each row corresponds to a rule. A rule can be copied into
\code{dplyr::filter} to filter the observations corresponding to a rule
}
\examples{
data("attrition", package = "rsample")
attrition <- tibble::as_tibble(attrition)
# lets predict monthly income
cubist_model <- 
  Cubist::cubist(
    x   = attrition \%>\% dplyr::select(-MonthlyIncome, -Attrition)
    , y = attrition \%>\% dplyr::select(MonthlyIncome) \%>\% unlist()
    )
summary(cubist_model)
tidy_rules(cubist_model)

cubist_model_commitees <- 
  Cubist::cubist(x   = attrition \%>\% dplyr::select(-MonthlyIncome, -Attrition)
                 , y = attrition \%>\% dplyr::select(MonthlyIncome) \%>\% unlist()
                 , committees = 7
                 )
summary(cubist_model_commitees)
tidy_rules(cubist_model_commitees)

# column names with spaces are handled with adding '`' quotes to it in the rules
ames <- AmesHousing::make_ames()

ames2 <- 
  ames \%>\%
  dplyr::rename(`Gr Liv Area` = Gr_Liv_Area) \%>\%
  dplyr::rename(`Gr Liv` = Latitude) \%>\% 
  dplyr::mutate(
    Overall_Qual = gsub("_", " ", as.character(Overall_Qual)),
    MS_SubClass = gsub("_", " ", as.character(MS_SubClass))
    )


colnames(ames2)

cb_mod <- 
  cubist(
    x = ames2 \%>\% dplyr::select(-Sale_Price),
    y = log10(ames2$Sale_Price),
    committees = 3
    ) 

tr <- tidy_rules(cb_mod)
tr
tr$rhs[[1]]
}
\author{
Srikanth KS, \email{sri.teach@gmail.com}
}
